requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.8
django-environ==0.8.1
flake8==6.0.0
iniconfig==1.1.1
mccabe==0.7.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.9
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
six==1.16.0
sqlparse==0.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
whitenoise==5.3.0
